FROMLIST: drm: panel: Add tianma nt36672a panel driver
v4: https://lkml.org/lkml/2020/8/11/712

Some Poco F1 phones have an LCD panel from Tianma, model nt36672a,
with a resolution of 1080x2246 that operates in DSI video mode.

Add the drm panel driver for it.

During testing, Benni Steini <bennisteinir@gmail.com> helped us fix
the reset sequence timing (from 10ms to 20ms), to get the bootanimation
to work on Android.

With current AOSP, we need to increase it to 200ms - this seems to be a
safe high value to avoid a white screen occasionally.

Signed-off-by: Sumit Semwal <sumit.semwal@linaro.org>
Cc: Benni Steini <bennisteinir@gmail.com>
[AmitP: Build fixes for v5.10]
Signed-off-by: Amit Pundir <amit.pundir@linaro.org>
Change-Id: I5fc418a90dc7a75cb6a01eee59d6de281075dea5